Just a Test w/ Sleep Function
If you want to try it out, only use the sleep function and see the delay between "Program Start" and "Program End." If you wish to try and download then upload to the actual FTC tool, you will have to edit the downloaded file and remove our onOpMode start block. That has not been changed to be consistent yet.