Windows in Auto can now be Resized
This includes a fix for the Unity View to be sized to the window.
Changed stuck button to howto button as I don't think the video was needed anymore. The video is still there but is a placeholder for the future howto video.
Made other changes to structure of some tabs.
Changes in local.html should only affect html and css.
Javascript changes were made only on the first script section, so the merge branch shouldn't be affected which I will mostly likely help out on next.